SD card: remove sd.c
This file was never in use and actually didn't even compile.
It didn't disturb so far because its whole content was wrapped
in #if 0. Recently it was fixed to at least compile, but code
quality was found to be poor enough to justify looking for
something written more carefully, which was found in FatFs
sample code.